machine-learning-2/tp-final/serving/download_models.py
from pathlib import Path
from google.cloud import storage


def download_files(bucket_name, bucket_dir):
    storage_client = storage.Client()
    bucket = storage_client.bucket(bucket_name)
    blobs = bucket.list_blobs(prefix=bucket_dir)
    for blob in blobs:
        if blob.name.endswith("/"):
            continue
        file_split = blob.name.split("/")
        directory = "/".join(file_split[0:-1])
        Path(directory).mkdir(parents=True, exist_ok=True)
        blob.download_to_filename(blob.name)


download_files("ml2-model-bucket", "mlruns/")
