machine-learning-2/tp-final/serving/requirements.txt
mlflow
sklearn
google-cloud-storage
psycopg2-binary